Move helper functions into Pipeline module
